changed float3 to double3 in segment intersection

+++ sources/structs.cpp
@@ -221,18 +221,18 @@ void Segment::ComputeBBox()
 
 bool Segment::Intersect(Ray& ray, RayIntersectionInfo& info)
 {
-	float3 s = v1 - v0;
-	float3 e = v0 - ray.GetOrigin();
-	float3 cross_dir_s = float3::cross(ray.GetDirection(), s);
-	float3 cross_e_dir = float3::cross(e, ray.GetDirection());
-	float x = float3::length(cross_dir_s);
-	float t = float3::length(float3::cross(e, s)) / x;
-	float u = float3::length(cross_e_dir) / x * sign(float3::dot(cross_dir_s, cross_e_dir));
+	double3 s = v1 - v0;
+	double3 e = v0 - ray.GetOrigin();
+	double3 cross_dir_s = double3::cross(ray.GetDirection(), s);
+	double3 cross_e_dir = double3::cross(e, ray.GetDirection());
+	double x = double3::length(cross_dir_s);
+	double t = double3::length(double3::cross(e, s)) / x;
+	double u = double3::length(cross_e_dir) / x * sign(double3::dot(cross_dir_s, cross_e_dir));
 	if (ray.GetMax() > t && t > ray.GetMin() && 1 > u && u >= 0.0f)
 	{
 		ray.SetMax(t);
-		float3 n = float3::cross(float3::cross(ray.GetDirection(), s), s);
-		info.SetNormal(float3::normalize(n));
+		double3 n = double3::cross(double3::cross(ray.GetDirection(), s), s);
+		info.SetNormal(double3::normalize(n));
 		info.AddClosestHit(t);
 		return true;
 	}
@@ -241,13 +241,13 @@ bool Segment::Intersect(Ray& ray, RayIntersectionInfo& info)
 
 bool Segment::AnyIntersect(Ray& ray)
 {
-	float3 s = v1 - v0;
-	float3 e = v0 - ray.GetOrigin();
-	float3 cross_dir_s = float3::cross(ray.GetDirection(), s);
-	float3 cross_e_dir = float3::cross(e, ray.GetDirection());
-	float x = float3::length(cross_dir_s);
-	float t = float3::length(float3::cross(e, s)) / x;
-	float u = float3::length(cross_e_dir) / x * sign(float3::dot(cross_dir_s, cross_e_dir));
+	double3 s = v1 - v0;
+	double3 e = v0 - ray.GetOrigin();
+	double3 cross_dir_s = double3::cross(ray.GetDirection(), s);
+	double3 cross_e_dir = double3::cross(e, ray.GetDirection());
+	double x = double3::length(cross_dir_s);
+	double t = double3::length(double3::cross(e, s)) / x;
+	double u = double3::length(cross_e_dir) / x * sign(double3::dot(cross_dir_s, cross_e_dir));
 	if (ray.GetMax() > t && t > ray.GetMin() && 1 > u && u >= 0.0f)
 	{
 		return true;
@@ -257,13 +257,13 @@ bool Segment::AnyIntersect(Ray& ray)
 
 bool Segment::AllIntersect(Ray& ray, RayIntersectionInfo& info)
 {
-	float3 s = v1 - v0;
-	float3 e = v0 - ray.GetOrigin();
-	float3 cross_dir_s = float3::cross(ray.GetDirection(), s);
-	float3 cross_e_dir = float3::cross(e, ray.GetDirection());
-	float x = float3::length(cross_dir_s);
-	float t = float3::length(float3::cross(e, s)) / x;
-	float u = float3::length(cross_e_dir) / x * sign(float3::dot(cross_dir_s, cross_e_dir));
+	double3 s = v1 - v0;
+	double3 e = v0 - ray.GetOrigin();
+	double3 cross_dir_s = double3::cross(ray.GetDirection(), s);
+	double3 cross_e_dir = double3::cross(e, ray.GetDirection());
+	double x = double3::length(cross_dir_s);
+	double t = double3::length(double3::cross(e, s)) / x;
+	double u = double3::length(cross_e_dir) / x * sign(double3::dot(cross_dir_s, cross_e_dir));
 	if (ray.GetMax() > t && t > ray.GetMin() && 1 > u && u >= 0.0f)
 	{
 		info.AddHit(t);
